Feature/create run example (#40)
* Intermediate progress

* Update run-example

Delaying while OpenAMASE opens appears to be important. It takes a while for OpenAMASE to actually open, and the user is supposed to click on the run button before OpenUxAS instances start. The five seconds offered by some examples is insufficient, and may not be guaranteed in any case.

We make the delay configurable, so that the user can remove or adjust it, as needed.

We also make sure that we kill off forked subprocesses and report if we're unable to do so.

* Rename the Ada Waterway Search example

Suggesting this naming convention so that we can create other examples that use the Ada UxAS more easily.

* Add two additional configs

Configs for Examples 03 and 05

* Update config for example 03

This should work out of the box on modern Linux.

* Add config-driven delay after OpenAMASE starts

This is important for the Ada example, although it may be related to the way in which we implemented it.

* Enhance run-example

- Fix error when OpenAMASE does not run
- Handle Ctrl-C gracefully (shut down all subprocesses)
- Allow a relative or absolute path to the example to be specified - this is nice, because then tab-complete works

* Add missing example configs

Also make sure that each ends with a newline (nice for `cat`).

* Use Test_SimulationTime service to sync with AMASE

Removes the need for any timeout to start UxAS instances.

* Fix example 05_AssignTasks

Adding the Test_SimulationTime service + a longer timeout on task assignment allows the example to run correctly.