Improve integration with anod. (#10)
Modify the makefile, RunLmcpGen and run-example scripts so that they
are aware of anod. They still work without anod, but if they find that
anod is under ~/bootstrap, they will each automatically ask anod to
configure the environment as appropriate for what they need to do.

1. For the makefile, this means getting the uxas --build-env from anod.
2. For RunLmcpGen, this means getting the lmcpgen --build-env from anod.
3. For run-example, things are even better than simply getting the envs
from anod.

Now, run-example will automatically use a locally-built uxas, if one
exists. It will also see if there’s a built OpenAMASE in ../OpenAMASE
(the expected place for a devel mode setup). At any point, it will fall
back on anod’s uxas or amase, warning that user that that’s what’s
happening, in the case of OpenAMASE (since I think that would be
surprising).

The upshot of this is the elimination of all of the anod printenv
commands.

Update the README to remove printenv commands.