Add an updated model for OpenUxAS
This model is currently incomplete and smaller than `uxas.lus`. However,
it is more realistic in the representation of routes and waypoints and
uses uninterpreted functions to represent deferred operations and
predicates over routes. Moreover, it provides a much more detailed
description of service behavior that is at a level suitable for handing
off to a language like SPARK for implementation.

The model fully proves with `jkind` (but is incompatible with `kind2`
because of the use of uninterpreted functions).