Enhancements to the model.
This includes significant changes to the theory of sequences to repair
identified flaws. In particular, the theory was overconstrained with
respect to repeated elements and underconstrained in the relationships
between get and first_index_of / last_index_of. This has been addressed.
The theory currently allows sequences with zero, one or two copies of a
given element - but no more than two copies.

The model has been enhanced to use an explicit notion of a bus for
describing the environment. This massively simplified the various
services and the uxas driver.

The model still does not prove in compositional mode.