Partial rework of WM
Attempts to use the same pattern for the WM as the ARV have failed - and
I don't understand why. Where the solvers were able to immediately prove
the property of interest for the ARV, they refuse to do so for the WM.
With painstaking breadcrumbing, I've managed to prove a subset of the
desired requirement - but only through the use of an unacceptable
trigger for my lemmas: a reference to an internal state variable.

This requires deeper investigation.