Further refactoring to improve the model.
The bus now enforces the constraint that only one service may respond
with a nonempty message at a time. Moreover, the bus is responsible for
the downselection to a single output message.

The notion of an output message at the uxas level is eliminated and all
properties are rephrased solely in terms of inputs.

All properties prove. As before, all properties should be enabled for
proof. Proof times appear quite variable - perhaps sensitivity to
initial conditions for the solvers? - but generally conclude in about 30
minutes on my machine.