Refactoring and working contract-based proofs.
Based on discussions with @daniel-larraz and using the most recent
commits to kind2, contract-based proof of uxas_props is now possible,
including when the task abstraction is included in the model.
Furthermore, with @daniel-larraz's help, a lemma has been found for the
task abstraction that allows P2 to prove very quickly.

In the current state, complete proof is possible with:

	kind2 uxas.lus --lus_main uxas_props --compositional true \
	--modular true

This takes about 60-90 seconds on my machine, depending on the run, and
uses kind2 version: `kind2 v1.4.0-36-g9bfcd460` (built from the github
sources).

Monolithic proof appears to remain out of reach; it is expected that
monolithic proof will be decreasingly successful as the model complexity
increases.

In addition to restoring proof, this commit:

* refactors contracts using the contract-import syntax, which is nicer
  to read

* removes properties redundant because they are included in contracts

* uses the node-import approach to create a partially defined but
  constrained node (see plan_builder.lus)

* renames task_response to task_complete

* disables all antecedent checks (for speed and clarity of results)