Fix errors for RunLmcpGen.py (#30)
RunLmcpGen.py no longer worked; the paths returned by the `uxas.paths`
python module were not correct; they were relative to the parent
directory of the OpenUxAS repo, not to the OpenUxAS repo itself.

Most likely, this occurred because of some change in the way the module
was installed in the python venv. To make this more robust, the paths
are now computed different: they look explicitly for the parent of the
.vpython directory itself.

Additionally, there was a possibility that the computation of the
`base_env` in RunLmcpGen.py would fail: the loop indexing requires that
the length of the array mod 2 be zero. This is now encoded explicitly.