Improve run_example.py
Recent updates to linters revealed a number of potential code-quality
issues.

Most were related to the use of f-strings in loggers, which is
discouraged because the string interpolation is greedy (happens even if
that logging level is disabled) and because it makes the job of log
aggregators more difficult (unlikely to matter to us).

There was an instance of use of a "global" that was not intended to be
such. This was spotted by moving the core script-like functionality into
a function, which avoided adding names to the global namespace.

The script is also renamed to conform to python module-naming
conventions.

All warnings and errors are resolved, except the overly broad catching
of Exception, which is explicitly silenced.

Add a workflow for running the HelloWorld example in CI as a simple
sanity check on run-example functionality.