Update infrastructure and run_lmcgen
This commit makes a number of improvements.

1. Update paths.sh to allow env overrides

Previously, paths.sh set path variables without checking to see if they
had already been set by the user. This is now fixed: if the variable is
already set, that value will be used. Otherwise, the default value will
be used.

This will allow the user to more easily steer the behavior of the
infrastructure (although care must be taken to ensure that values
provided in the environment make sense).

2. Update paths.py with docstrings and to match paths.sh

3. Update run_example.py

Some minor problems in run_example are resolved / cleaned up. There is
likely more work to be done here; this is a complex file.

4. Update run_lmcpgen.py

Significant improvements to usability through better handling of paths
and better error reporting.

run_lmcpgen will now run LmcpGen from:
  1. command-line argument - building if needed
  2. explicit environment variable (LMCP_DEVEL_DIR) - building if needed
  3. directory spec'd in repositories.yaml - building if needed
  4. sanbox - failing if not already build

run_lmcpgen will build for OpenAMASE if -a is given, building in:
  1. command-line argument
  2. explicit environment variable (AMASE_DEVEL_DIR)
  3. directory spec'd in repositories.yaml

OpenAMASE in the sandbox is not used, as changes would be overwritten on
next anod build.