Updates to support Ubuntu-22 and GCC11
GCC 11 changes the header file for `std::numerical_limits`; this fix
addresses that change.

gcov with GCC 11 drops `__gcov_flush` and recommends the use of the
`__gcov_dump` instead.

The `bootstrap.patch` is a long-unused file.

The GitHub workflows are all updated to include Ubuntu-22 in the matrix
builds. We should remove Ubuntu-18 in a future issue.