Adapt Ada code and take advantage of new SPARK features (#74)
* Adapt Ada code and take advantage of new SPARK features

This patch makes the Ada code compatible with the changes done
in GNAT and SPARK. It takes advantage of new features such as
logical equality, both in containers and more generally in
SPARK code, to remove lots of wrappers that were used to
reduce context for provers to prove invariants on the
service state.

* Replace bounded formal containers by unbounded formal containers

[changelog]
* src/ada/src/services/
Replace bounded formal containers by unbounded formal containers in
Route aggregator and Assignment Tree Branch Bound. Fix proofs.

* src/ada/proof/sessions
Update session files.

* tests/proofs/
Update tests outputs.

* WIP Alire integration

Install working; need to make use of it in anod for printenv.

* Leverage the use of alr in `ensure_gnat`

This allows us to use gnat and gnatprove, after they've been installed
by Alire.

* Add XMLAda and fix auto-install.

Last commit broke auto-install of GNAT; this fixes it.

Also adds XMLAda from alr. It appears XMLAda was removed from the
alr-provided GNAT.

This patch is still broken, however, because of possible gprbuild
issues.

* Build XMLAda with alr.

* Fix spurious spaces in anod script

* Install xmlada via anod, rather than Alire.

This works around a problem with the Alire version not being
relocatable.

* Use infrastructure install for gnat

... rather than the GitHub action.

* Update Alire support for 13.2.1

Note: this commit is in an intermediate state, as we're waiting for
resolution of an Alire config-directory-handling error.

* Updates to enable successful build with 13.2.0

* Add commands to work around Alire bug.

* Add support for adding GNAT FSF to the path.

This will happen when we do `anod printenv uxas-ada --build-env`

* Use the nightly build of Alire for local config

The nightly build of Alire has the fix to use local configs for
toolchain selection. This removes the need for the workaround we put in
place on a prior commit.

Once there's a new release of Alire with this fix, we'll update again to
use that version, for stability.

* Update code to get rid of linker errors using sparklib.gpr.
Update session files.
Update proof tests outputs.

---------

Co-authored-by: M. Anthony Aiello <aiello@adacore.com>